+++ src/core/main/video-detail.js
@@ -1,4 +1,4 @@
-import Modal from "../../components/modal.styled";
+import Modal from "../../components/modal";
 import React, { useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { useDispatch, useSelector } from "react-redux";
