+++ src/core/movies/section/index.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import MoviesSection from "./movies-section";
+import MoviesSection from "./section.component";
 
 export default styled(MoviesSection)`
   background-color: #343a40;
