+++ src/layouts/brand/index.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import Logo from "./logo";
+import Logo from "./logo.component";
 
 export default styled(Logo)`
   .header__logo {
