+++ src/layouts/footer/index.js
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import Footer from "./footer";
+import Footer from "./footer.component";
 
 export default styled(Footer)`
   background-color: #212529;
